requirements.txt
pyqt6
google-api-python-client
google-auth-httplib2 
google-auth-oauthlib